requirements.txt
Django==2.1.4
django-crispy-forms==1.7.2
django-decouple==2.1
django-filter==2.0.0
django-widgets-improved==1.5.0
djangorestframework==3.9.0
Markdown==3.0.1
psycopg2==2.7.6.1
pytz==2018.7
